Refactor: move esi.rs into own directory and separate methods into multiple files.
